requirements.txt
# Core dependencies
pyyaml>=6.0
python-dotenv>=1.0.0

# Fireworks AI SDK
fireworks-ai>=0.9.0

# Data processing
jsonlines>=4.0.0
